remove number of shards
from google support:
They did confirm that if you remove the sharding specification on the WriteToFiles transform, the files will be written to directly bypassing the grouping operation within the WriteToFiles transform. This will be faster as the records will be read from Bigtable and written to GCS directly.

This will reduce the cost as the number of bytes shuffled decreases and should help in job runtime as-well since CPU won't be used writing to and reading from the shuffle (aka GroupByKey).

The downside to this is that this won't allow you have control on the number of files written.